Gracefully return null if the service is null
If the plugin isn't activated yet, the analyzers will still try to check
if they can analyze the program. But if the plugin isn't activated the
service returned will be `null` and the crash will result in multiple
annoying popups.
This gracefully handles this, by propagating the null return value.